requirements.txt
attrs==23.2.0 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-asset-awscli-v1==2.2.202 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-asset-kubectl-v20==2.1.2 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-asset-node-proxy-agent-v6==2.0.3 ; python_version >= "3.12" and python_version < "4.0"
aws-cdk-lib==2.138.0 ; python_version >= "3.12" and python_version < "4.0"
cattrs==23.2.3 ; python_version >= "3.12" and python_version < "4.0"
constructs==10.3.0 ; python_version >= "3.12" and python_version < "4.0"
importlib-resources==6.4.0 ; python_version >= "3.12" and python_version < "4.0"
jsii==1.97.0 ; python_version >= "3.12" and python_version < "4.0"
publication==0.0.3 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
typeguard==2.13.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.12" and python_version < "4.0"
